+++ packages/gatsby-source-wordpress-experimental/src/steps/ingest-remote-schema/build-queries-from-introspection/build-query-on-field-name.js
@@ -1,4 +1,6 @@
 import compress from "graphql-query-compress"
+import store from "~/store"
+import { findTypeName } from "~/steps/create-schema-customization/helpers"
 
 export const buildNodesQueryOnFieldName = ({
   fields,
@@ -47,13 +49,17 @@ export const buildSelectionSet = fields => {
     return ``
   }
 
+  const {
+    remoteSchema: { typeMap },
+  } = store.getState()
+
   return fields
     .map(field => {
       if (typeof field === `string`) {
         return field
       }
 
-      let { fieldName, variables, fields, inlineFragments } = field
+      let { fieldName, variables, fields, inlineFragments, fieldType } = field
 
       // @todo instead of checking for a nodes field, include the field type here
       // and check for input args instead. Maybe some kind of input args API or something would be helpful
@@ -75,6 +81,15 @@ export const buildSelectionSet = fields => {
           }
         `
       } else if (fieldName) {
+        const fullFieldType = typeMap.get(findTypeName(fieldType))
+
+        // if this field has subfields but we didn't build a selection set for it
+        // we shouldn't fetch this field. This can happen when we have self referencing types that are limited by the schema.circularQueryLimit plugin option.
+        // @todo the above should be fixed in recursively-transform-fields.js instead of here. recursion is hard :p
+        if (fullFieldType.fields) {
+          return null
+        }
+
         return fieldName
       }